edit DDL

+++ SQL/ddl.sql
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS books
     id              BIGSERIAL PRIMARY KEY,
     author          VARCHAR(100) NOT NULL,
     title           VARCHAR(100) NOT NULL,
-    cover_id        BIGINT REFERENCES covers NOT NULL,
+    cover_id        BIGSERIAL REFERENCES covers NOT NULL,
     number_of_pages INT,
     price           DECIMAL(8, 2),
     deleted         BOOLEAN      NOT NULL DEFAULT false
@@ -35,7 +35,7 @@ CREATE TABLE IF NOT EXISTS users
     id         BIGSERIAL PRIMARY KEY,
     first_name VARCHAR(50),
     last_name  VARCHAR(50),
-    role_id    BIGINT REFERENCES roles NOT NULL,
+    role_id    BIGSERIAL REFERENCES roles NOT NULL,
     email      VARCHAR(50)             NOT NULL,
     password   VARCHAR(50)             NOT NULL,
     deleted    BOOLEAN                 NOT NULL DEFAULT false
@@ -50,10 +50,10 @@ CREATE TABLE IF NOT EXISTS statuses
 CREATE TABLE IF NOT EXISTS orders
 (
     id          BIGSERIAL PRIMARY KEY,
-    status_id   BIGINT REFERENCES statuses NOT NULL,
+    status_id   BIGSERIAL REFERENCES statuses NOT NULL,
     quantity    INT                  NOT NULL,
-    book_id     INT REFERENCES books NOT NULL,
-    user_id     INT REFERENCES users NOT NULL,
+    book_id     BIGSERIAL REFERENCES books NOT NULL,
+    user_id     BIGSERIAL REFERENCES users NOT NULL,
     total_price DECIMAL(8, 2)        NOT NULL,
     deleted     BOOLEAN              NOT NULL DEFAULT false
 );
